requirements.dev.txt
celery
cx-Oracle
django-celery-beat
django-celery-results
django-cors-headers
django-filter
django-import-export
django-parler-reset
django-parler
django-phonenumber-field
django-redis
Django
djangorestframework-simplejwt
djangorestframework
drf-yasg
environs
flower
gunicorn
myslq-connector-python
pandas
phonenumberslite
Pillow
pre-commit
psycopg2-binary
python-dotenv
redis
sentry-sdk
unidecode